ml-service/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
